Cleanup: move DBusManager methods to cpp file
